requirements.txt
openai
request